Move editor to same package as viewer Also testing production export of application
