+++ .idea/runConfigurations/Main.xml
@@ -2,11 +2,11 @@
   <configuration default="false" name="Main" type="Application" factoryName="Application" singleton="true">
     <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea">
       <pattern>
-        <option name="PATTERN" value="editor.*" />
+        <option name="PATTERN" value="ui.*" />
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
-    <option name="MAIN_CLASS_NAME" value="viewer.Main" />
+    <option name="MAIN_CLASS_NAME" value="ui.Main" />
     <option name="VM_PARAMETERS" value="" />
     <option name="PROGRAM_PARAMETERS" value="" />
     <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$" />
