+++ src/algorithm/Algorithm.java
@@ -4,8 +4,8 @@ import javafx.geometry.Point2D;
 
 import java.util.ArrayList;
 
-import static viewer.Controller.boundary;
-import static viewer.Controller.robot;
+import static ui.Controller.boundary;
+import static ui.Controller.robot;
 
 public class Algorithm {
