+++ src/ui/Main.java
@@ -1,4 +1,4 @@
-package viewer;
+package ui;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
